refactor(bar): moved widgets to new folder
This commit moved all the widgets to a new folder in order to organize the code while hoping to make it easier to find and add a widget.
