+++ komorebi-bar/src/bar.rs
@@ -4,15 +4,15 @@ use crate::config::KomobarTheme;
 use crate::config::MonitorConfigOrIndex;
 use crate::config::Position;
 use crate::config::PositionConfig;
-use crate::komorebi::Komorebi;
-use crate::komorebi::KomorebiNotificationState;
 use crate::process_hwnd;
 use crate::render::Color32Ext;
 use crate::render::Grouping;
 use crate::render::RenderConfig;
 use crate::render::RenderExt;
-use crate::widget::BarWidget;
-use crate::widget::WidgetConfig;
+use crate::widgets::komorebi::Komorebi;
+use crate::widgets::komorebi::KomorebiNotificationState;
+use crate::widgets::widget::BarWidget;
+use crate::widgets::widget::WidgetConfig;
 use crate::KomorebiEvent;
 use crate::BAR_HEIGHT;
 use crate::DEFAULT_PADDING;
