+++ komorebi-bar/src/widgets/date.rs
@@ -1,7 +1,7 @@
 use crate::config::LabelPrefix;
 use crate::render::RenderConfig;
 use crate::selected_frame::SelectableFrame;
-use crate::widget::BarWidget;
+use crate::widgets::widget::BarWidget;
 use chrono::Local;
 use chrono_tz::Tz;
 use eframe::egui::text::LayoutJob;
