+++ komorebi-bar/src/widgets/keyboard.rs
@@ -1,6 +1,6 @@
 use crate::config::LabelPrefix;
 use crate::render::RenderConfig;
-use crate::widget::BarWidget;
+use crate::widgets::widget::BarWidget;
 use eframe::egui::text::LayoutJob;
 use eframe::egui::Align;
 use eframe::egui::Context;
