+++ komorebi-bar/src/widgets/komorebi.rs
@@ -2,12 +2,12 @@ use crate::bar::apply_theme;
 use crate::config::DisplayFormat;
 use crate::config::KomobarTheme;
 use crate::config::WorkspacesDisplayFormat;
-use crate::komorebi_layout::KomorebiLayout;
 use crate::render::Grouping;
 use crate::render::RenderConfig;
 use crate::selected_frame::SelectableFrame;
 use crate::ui::CustomUi;
-use crate::widget::BarWidget;
+use crate::widgets::komorebi_layout::KomorebiLayout;
+use crate::widgets::widget::BarWidget;
 use crate::ICON_CACHE;
 use crate::MAX_LABEL_WIDTH;
 use crate::MONITOR_INDEX;
