+++ komorebi-bar/src/widgets/mod.rs
@@ -1,0 +1,13 @@
+pub mod battery;
+pub mod cpu;
+pub mod date;
+pub mod keyboard;
+pub mod komorebi;
+mod komorebi_layout;
+pub mod media;
+pub mod memory;
+pub mod network;
+pub mod storage;
+pub mod time;
+pub mod update;
+pub mod widget;
