+++ komorebi-bar/src/widgets/network.rs
@@ -1,7 +1,7 @@
 use crate::config::LabelPrefix;
 use crate::render::RenderConfig;
 use crate::selected_frame::SelectableFrame;
-use crate::widget::BarWidget;
+use crate::widgets::widget::BarWidget;
 use eframe::egui::text::LayoutJob;
 use eframe::egui::Align;
 use eframe::egui::Context;
