+++ komorebi-bar/src/widgets/widget.rs
@@ -1,26 +1,26 @@
-use crate::battery::Battery;
-use crate::battery::BatteryConfig;
-use crate::cpu::Cpu;
-use crate::cpu::CpuConfig;
-use crate::date::Date;
-use crate::date::DateConfig;
-use crate::keyboard::Keyboard;
-use crate::keyboard::KeyboardConfig;
-use crate::komorebi::Komorebi;
-use crate::komorebi::KomorebiConfig;
-use crate::media::Media;
-use crate::media::MediaConfig;
-use crate::memory::Memory;
-use crate::memory::MemoryConfig;
-use crate::network::Network;
-use crate::network::NetworkConfig;
 use crate::render::RenderConfig;
-use crate::storage::Storage;
-use crate::storage::StorageConfig;
-use crate::time::Time;
-use crate::time::TimeConfig;
-use crate::update::Update;
-use crate::update::UpdateConfig;
+use crate::widgets::battery::Battery;
+use crate::widgets::battery::BatteryConfig;
+use crate::widgets::cpu::Cpu;
+use crate::widgets::cpu::CpuConfig;
+use crate::widgets::date::Date;
+use crate::widgets::date::DateConfig;
+use crate::widgets::keyboard::Keyboard;
+use crate::widgets::keyboard::KeyboardConfig;
+use crate::widgets::komorebi::Komorebi;
+use crate::widgets::komorebi::KomorebiConfig;
+use crate::widgets::media::Media;
+use crate::widgets::media::MediaConfig;
+use crate::widgets::memory::Memory;
+use crate::widgets::memory::MemoryConfig;
+use crate::widgets::network::Network;
+use crate::widgets::network::NetworkConfig;
+use crate::widgets::storage::Storage;
+use crate::widgets::storage::StorageConfig;
+use crate::widgets::time::Time;
+use crate::widgets::time::TimeConfig;
+use crate::widgets::update::Update;
+use crate::widgets::update::UpdateConfig;
 use eframe::egui::Context;
 use eframe::egui::Ui;
 use serde::Deserialize;
